add tag and repo

+++ src/main/java/entities/User.java
@@ -6,14 +6,14 @@ import lombok.Setter;
 @Getter
 @Setter
 public class User {
-    private int id;
+    private long id;
     private String displayName;
     private String email;
     private String username;
     private String password;
     private String bio;
 
-    public User(int id, String displayName, String email, String username, String password, String bio) {
+    public User(long id, String displayName, String email, String username, String password, String bio) {
         this.id = id;
         this.displayName = displayName;
         this.email = email;
